move domain classes in separate module
