--- src/main/java/ro/msg/learning/shop/model/ProductCategory.java
@@ -1,18 +1,0 @@
-package ro.msg.learning.shop.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@EqualsAndHashCode(callSuper = true)
-public class ProductCategory extends GenericEntity {
-    private String name;
-    private String description;
-}
