--- src/main/java/ro/msg/learning/shop/model/Stock.java
@@ -1,24 +1,0 @@
-package ro.msg.learning.shop.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.ManyToOne;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@EqualsAndHashCode(callSuper = true)
-//TODO: composite key?
-public class Stock extends GenericEntity {
-    @ManyToOne(fetch = FetchType.EAGER)
-    private Product product;
-    @ManyToOne(fetch = FetchType.EAGER)
-    private Location location;
-    private Integer quantity;
-}
