+++ src/main/java/ro/msg/learning/shop/model/domain/ProductCategory.java
@@ -1,4 +1,4 @@
-package ro.msg.learning.shop.model;
+package ro.msg.learning.shop.model.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -12,6 +12,7 @@ import javax.persistence.Entity;
 @AllArgsConstructor
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class Supplier extends GenericEntity {
+public class ProductCategory extends GenericEntity {
     private String name;
+    private String description;
 }
