+++ src/main/java/ro/msg/learning/shop/model/domain/Stock.java
@@ -1,28 +1,23 @@
-package ro.msg.learning.shop.model;
+package ro.msg.learning.shop.model.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.ManyToOne;
-import java.time.LocalDateTime;
 
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 @EqualsAndHashCode(callSuper = true)
-//TODO: change fetch type to lazy eventually
-public class Order extends GenericEntity {
+public class Stock extends GenericEntity {
     @ManyToOne(fetch = FetchType.EAGER)
-    private Location shippedFrom;
+    private Product product;
     @ManyToOne(fetch = FetchType.EAGER)
-    private Customer customer;
-    private LocalDateTime createdAt;
-    @Embedded
-    private Address address;
+    private Location location;
+    private Integer quantity;
 }
