+++ src/main/java/ro/msg/learning/shop/persistence/CustomerRepository.java
@@ -2,7 +2,7 @@ package ro.msg.learning.shop.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ro.msg.learning.shop.model.Customer;
+import ro.msg.learning.shop.model.domain.Customer;
 
 @Repository
 public interface CustomerRepository extends JpaRepository<Customer, Integer> {
