+++ src/main/java/ro/msg/learning/shop/persistence/LocationRepository.java
@@ -2,7 +2,7 @@ package ro.msg.learning.shop.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ro.msg.learning.shop.model.Location;
+import ro.msg.learning.shop.model.domain.Location;
 
 @Repository
 public interface LocationRepository extends JpaRepository<Location, Integer> {
