+++ src/main/java/ro/msg/learning/shop/persistence/OrderRepository.java
@@ -2,7 +2,7 @@ package ro.msg.learning.shop.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ro.msg.learning.shop.model.Order;
+import ro.msg.learning.shop.model.domain.Order;
 
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Integer> {
